apenas funcionários responsáveis pela os podem exclui-lá

+++ Projeto/apps/views.py
@@ -8,6 +8,8 @@ from .utils import filtrar_ordens
 from .models import *
 from django.contrib import auth
 from django.http import HttpResponse
+from django.http import HttpResponseForbidden
+
 
 #VIEWS DE LOGIN
 def login_view(request):
@@ -261,18 +263,22 @@ def servicos(request):
         return render(request, 'apps/servicos.html', {'funcionario': 1, 'ordens_servico': ordens_servico})
 
 @login_required
-def listar_os(request): #listar todas as os feitas 
+def listar_os(request):
     user = request.user
-    usuario = Perfil.objects.get(username=user)
+    usuario = Perfil.objects.get(username=user.username)
 
     if usuario.funcionario == 0:
         return redirect(login)
     else:
-        if request.user.is_anonymous:
-            return redirect(login)
-        else:
-            ordens = OrdemServico.objects.all()
-            return render(request, 'apps/listar_os.html', {'funcionario': 1, 'ordens': ordens})
+        ordens = OrdemServico.objects.all()
+        ordens, status, data_criacao = filtrar_ordens(request, ordens)
+        return render(request, 'apps/listar_os.html', {
+            'funcionario': 1,
+            'ordens': ordens,
+            'status': status,
+            'data_criacao': data_criacao,
+            'usuario': usuario  # Passa o perfil do usuário logado para o contexto
+        })
 
 @login_required
 def editar_os(request, os_id):
@@ -354,9 +360,19 @@ def detalhes_os(request, os_id):
 @login_required
 def excluir_os(request, pk):
     os = get_object_or_404(OrdemServico, pk=pk)
+
+    # Verifica se há um funcionário responsável pela OS
+    if not os.funcionario_responsavel:
+        return HttpResponseForbidden("Não é possível excluir uma OS sem um funcionário responsável.")
+
+    # Verifica se o usuário logado é o funcionário responsável pela OS
+    if os.funcionario_responsavel.username != request.user.username:
+        return HttpResponseForbidden("Você não tem permissão para excluir esta OS.")
+
     if request.method == 'POST':
         os.delete()
         return redirect('listar_os')  # Redireciona para a lista de OS após a exclusão
+
     return render(request, 'excluir_os.html', {'os': os})
 
 @login_required
